split multi logger into another file
